daemon: remove pidfile and socket only when created by this instance
It may happen that dom0 QubesDB instance was started again for the same
domain - for example when restarting domain. Old instance will
eventually notice its domain disappearance, but with noticeable delay -
10s in worst case.
Do not remove pidfile and socket in such a case - now those belongs to
the new instance.

There is still slight race condition - between stat and unlink call. But
this is very unlikely to cause problems.

Fixes QubesOS/qubes-issues#1694